feat:NavBar 구현 및 컴포넌트 분리
