+++ src/pages/homepage/components/NavBar.tsx
@@ -1,0 +1,61 @@
+/** @jsxImportSource @emotion/react */
+import { css } from '@emotion/react';
+import { IoCaretBack } from 'react-icons/io5';
+import { FaUser } from 'react-icons/fa';
+import theme from '../../../styles/theme';
+import logo from '../../assets/images/logo.png';
+const NavBar = () => {
+  return (
+    <nav css={navStyle}>
+      <button css={iconButtonStyle}>
+        <IoCaretBack size={50} color="#ffffffff" />
+      </button>
+      <button css={titleButtonStyle}>
+        <img src={logo} alt="무쉼사 로고" css={logoStyle} />
+      </button>
+      <button css={iconButtonStyle}>
+        <FaUser size={50} color="#ffffffff" />
+      </button>
+    </nav>
+  );
+};
+
+export default NavBar;
+
+const navStyle = css`
+  position: fixed;
+  top: 0;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 100%;
+  max-width: 720px;
+  height: ${theme.spacing.spacing16};
+  padding: 0 ${theme.spacing.spacing4};
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  border-bottom: 1px solid ${theme.colors.gray200};
+  background-color: #000000ff;
+  z-index: 1000;
+`;
+
+const iconButtonStyle = css`
+  background: none;
+  border: none;
+  padding: 0;
+  margin: 0;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+`;
+
+const titleButtonStyle = css`
+  background: none;
+  border: none;
+  padding: 0;
+  cursor: pointer;
+`;
+
+const logoStyle = css`
+  height: ${theme.spacing.spacing16};
+`;
